fastapi-app/requirements.txt
fastapi
uvicorn
pandas
pymysql
python-dotenv
SQLAlchemy>=2.0